move response builder to controller
